Updates to CheckIn/Out
I made changes to checkInScreen:
1) Validated the text boxes for correct data
2) Check to make sure they all have data

Changes to checkOutScreen:
1) Forced it to put focus on the text box so that they don't have to
click and if they use a scanner it will put that value in there. At
least it should, if there is a text box selected anywhere, an attached
scanner should just upload a string. We will have to test that
obviously, but that's how they have worked in my experience.